tests: Avoid execvp() aborting with --pdb-trace
Our `execvp()` wrapper uses a subprocess if an env var is set.
And SnoCliRunner sets the env var when invoking a sno command.
But using `--pdb-trace` caused the SnoCliRunner to not set the env var,
so the pytest process was being replaced when calling execvp().

THis fixes that by passing env through even when using `--pdb-trace`