gc: Avoid race condition in test teardown
By default, `git gc --auto` returns immediately and actually runs in
the background.

This can cause race conditions with the tests where the test cleanup
uses `shutil.rmtree()` to remove the repo git gc is still running in,
and some OSError is raised.

This change forces gc to run syncronously in the test environment.